services/elpyfi-ai/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
anthropic==0.40.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
rich==13.9.4
requests==2.32.3
aiohttp==3.11.10
matplotlib==3.9.3
psycopg2-binary==2.9.10